# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AfriLance freelance marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /proposals:
    get:
      operationId: listMyProposals
      tags: [proposals]
      summary: List proposals submitted by the current freelancer
      responses:
        "200":
          description: My proposals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProposalListResponse"
    post:
      operationId: createProposal
      tags: [proposals]
      summary: Submit a proposal for a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProposalBody"
      responses:
        "201":
          description: Proposal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
  /proposals/{id}:
    get:
      operationId: getProposal
      tags: [proposals]
      summary: Get a proposal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
        "404":
          description: Not found
    patch:
      operationId: updateProposalStatus
      tags: [proposals]
      summary: Update a proposal status (accept/reject)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProposalStatusBody"
      responses:
        "200":
          description: Updated proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
  /jobs/{jobId}/proposals:
    get:
      operationId: listJobProposals
      tags: [proposals]
      summary: List proposals for a specific job (client view)
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job proposals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProposalListResponse"

  # Messages
...
components:
  schemas:
...
    Proposal:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        freelancerId:
          type: integer
        coverLetter:
          type: string
        bidAmount:
          type: number
        status:
          type: string
        freelancerName:
          type: ["string", "null"]
        freelancerAvatarUrl:
          type: ["string", "null"]
        freelancerLocation:
          type: ["string", "null"]
        freelancerSkills:
          type: array
          items:
            type: string
        freelancerAvgRating:
          type: ["number", "null"]
        jobTitle:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - jobId
        - freelancerId
        - coverLetter
        - bidAmount
        - status
        - freelancerSkills
        - createdAt
        - updatedAt
    CreateProposalBody:
      type: object
      properties:
        jobId:
          type: integer
        coverLetter:
          type: string
        bidAmount:
          type: number
      required:
        - jobId
        - coverLetter
        - bidAmount
    UpdateProposalStatusBody:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ProposalListResponse:
      type: object
      properties:
        proposals:
          type: array
          items:
            $ref: "#/components/schemas/Proposal"
        total:
          type: integer
      required:
        - proposals
        - total
...