# lib/api-zod/src/generated/types/createProposalBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AfriLance freelance marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProposalBody {
  jobId: number;
  coverLetter: string;
  bidAmount: number;
}