# lib/api-zod/src/generated/types/updateProposalStatusBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AfriLance freelance marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProposalStatusBody {
  status: string;
}